+++ node-api/src/test/kotlin/net/corda/nodeapi/internal/serialization/ContractAttachmentSerializerTest.kt
@@ -7,6 +7,7 @@ import net.corda.testing.contracts.DummyContract
 import net.corda.testing.core.SerializationEnvironmentRule
 import net.corda.testing.internal.rigorousMock
 import net.corda.testing.node.MockServices
+import org.apache.commons.lang.ArrayUtils.EMPTY_BYTE_ARRAY
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.Assert.assertArrayEquals
@@ -35,7 +36,7 @@ class ContractAttachmentSerializerTest {
 
     @Test
     fun `write contract attachment and read it back`() {
-        val contractAttachment = ContractAttachment(GeneratedAttachment(ByteArray(0)), DummyContract.PROGRAM_ID)
+        val contractAttachment = ContractAttachment(GeneratedAttachment(EMPTY_BYTE_ARRAY), DummyContract.PROGRAM_ID)
         // no token context so will serialize the whole attachment
         val serialized = contractAttachment.serialize(factory, context)
         val deserialized = serialized.deserialize(factory, context)
@@ -88,8 +89,7 @@ class ContractAttachmentSerializerTest {
 
     @Test
     fun `check attachment in deserialize is lazy loaded when using token context`() {
-        val attachment = GeneratedAttachment(ByteArray(0))
-
+        val attachment = GeneratedAttachment(EMPTY_BYTE_ARRAY)
         // don't importAttachment in mockService
 
         val contractAttachment = ContractAttachment(attachment, DummyContract.PROGRAM_ID)