+++ node-api/src/test/kotlin/net/corda/nodeapi/internal/serialization/MapsSerializationTest.kt
@@ -6,8 +6,8 @@ import net.corda.core.identity.CordaX500Name
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.serialization.deserialize
 import net.corda.core.serialization.serialize
-import net.corda.node.services.statemachine.SessionData
-import net.corda.nodeapi.internal.serialization.kryo.KryoHeaderV0_1
+import net.corda.node.services.statemachine.DataSessionMessage
+import net.corda.nodeapi.internal.serialization.kryo.kryoMagic
 import net.corda.testing.core.SerializationEnvironmentRule
 import net.corda.testing.internal.amqpSpecific
 import net.corda.testing.internal.kryoSpecific
@@ -41,7 +41,7 @@ class MapsSerializationTest {
 
     @Test
     fun `check list can be serialized as part of SessionData`() {
-        val sessionData = SessionData(123, smallMap.serialize())
+        val sessionData = DataSessionMessage(smallMap.serialize())
         assertEqualAfterRoundTripSerialization(sessionData)
         assertEquals(smallMap, sessionData.payload.deserialize())
     }
@@ -78,7 +78,7 @@ class MapsSerializationTest {
             val nameID = 0
             val serializedForm = emptyMap<Int, Int>().serialize()
             val output = ByteArrayOutputStream().apply {
-                write(KryoHeaderV0_1.bytes)
+                kryoMagic.writeTo(this)
                 write(DefaultClassResolver.NAME + 2)
                 write(nameID)
                 write(javaEmptyMapClass.name.toAscii())