+++ node/src/test/kotlin/net/corda/node/messaging/InMemoryMessagingTests.kt
@@ -3,7 +3,6 @@ package net.corda.node.messaging
 import net.corda.core.messaging.AllPossibleRecipients
 import net.corda.node.services.messaging.Message
 import net.corda.node.services.messaging.TopicStringValidator
-import net.corda.node.services.messaging.createMessage
 import net.corda.testing.internal.rigorousMock
 import net.corda.testing.node.MockNetwork
 import org.junit.After
@@ -50,10 +49,10 @@ class InMemoryMessagingTests {
 
         val bits = "test-content".toByteArray()
         var finalDelivery: Message? = null
-        node2.network.addMessageHandler { msg, _ ->
+        node2.network.addMessageHandler("test.topic") { msg, _ ->
             node2.network.send(msg, node3.network.myAddress)
         }
-        node3.network.addMessageHandler { msg, _ ->
+        node3.network.addMessageHandler("test.topic") { msg, _ ->
             finalDelivery = msg
         }
 
@@ -62,7 +61,7 @@ class InMemoryMessagingTests {
 
         mockNet.runNetwork(rounds = 1)
 
-        assertTrue(Arrays.equals(finalDelivery!!.data, bits))
+        assertTrue(Arrays.equals(finalDelivery!!.data.bytes, bits))
     }
 
     @Test
@@ -74,7 +73,7 @@ class InMemoryMessagingTests {
         val bits = "test-content".toByteArray()
 
         var counter = 0
-        listOf(node1, node2, node3).forEach { it.network.addMessageHandler { _, _ -> counter++ } }
+        listOf(node1, node2, node3).forEach { it.network.addMessageHandler("test.topic") { _, _ -> counter++ } }
         node1.network.send(node2.network.createMessage("test.topic", data = bits), rigorousMock<AllPossibleRecipients>())
         mockNet.runNetwork(rounds = 1)
         assertEquals(3, counter)
@@ -93,9 +92,8 @@ class InMemoryMessagingTests {
         node1.network.addMessageHandler("valid_message") { _, _ ->
             received++
         }
-
-        val invalidMessage = node2.network.createMessage("invalid_message", data = ByteArray(0))
-        val validMessage = node2.network.createMessage("valid_message", data = ByteArray(0))
+        val invalidMessage = node2.network.createMessage("invalid_message", data = ByteArray(1))
+        val validMessage = node2.network.createMessage("valid_message", data = ByteArray(1))
         node2.network.send(invalidMessage, node1.network.myAddress)
         mockNet.runNetwork()
         assertEquals(0, received)
@@ -106,8 +104,8 @@ class InMemoryMessagingTests {
 
         // Here's the core of the test; previously the unhandled message would cause runNetwork() to abort early, so
         // this would fail. Make fresh messages to stop duplicate uniqueMessageId causing drops
-        val invalidMessage2 = node2.network.createMessage("invalid_message", data = ByteArray(0))
-        val validMessage2 = node2.network.createMessage("valid_message", data = ByteArray(0))
+        val invalidMessage2 = node2.network.createMessage("invalid_message", data = ByteArray(1))
+        val validMessage2 = node2.network.createMessage("valid_message", data = ByteArray(1))
         node2.network.send(invalidMessage2, node1.network.myAddress)
         node2.network.send(validMessage2, node1.network.myAddress)
         mockNet.runNetwork()